% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{am_variables}
\alias{am_variables}
\title{The variables that go into the common data structure.}
\usage{
am_variables(varnames = "all")
}
\arguments{
\item{varnames}{if "all" (default), returns all values. Otherwise returns
a subset whose names corresponds to the values given.}
}
\value{
a list as described above.
}
\description{
The names of the list elements are the variable names, and the values are
either the legal choices or an empty vector of the correct data type.
}
