% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{amu_dummy_data}
\alias{amu_dummy_data}
\title{Generate some random artificial AMU data adhering to the variable rules
defined by the function \code{am_variables}. ATC codes are fetched from the
"atc.csv" file in this package, whose source is the Swedish Board
of Agriculture. The EntryID is merely an index of the row number.}
\usage{
amu_dummy_data(n_rows, start_date, end_date)
}
\arguments{
\item{n_rows}{the number of rows in the dataset}

\item{start_date}{The earliest possible date of transaction}

\item{end_date}{The latest possible date of transaction.}
}
\value{
a data.frame with the dummy data
}
\description{
NOTE! This function takes quite some time to run for large values of
\code{n_rows}.
}
