% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_spatial_data}
\alias{get_spatial_data}
\title{Extract spatial data for a given list of European countries}
\usage{
get_spatial_data(country_codes)
}
\arguments{
\item{country_codes}{official NUTS country codes for the countries to
retrieve data from.}
}
\value{
an \code{sf} object with the desired geospatial data.
}
\description{
Extract spatial data for a given list of European countries
}
