% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\docType{data}
\name{herdsize}
\alias{herdsize}
\title{Herdsize, indegree and outdegree per ppn and time}
\format{
A data frame with 2090063 rows and 7 variables:
\describe{
\item{ppn}{test}
\item{time}{..}
\item{size}{..}
\item{region}{..}
\item{inDegree}{..}
\item{outDegree}{..}
\item{storleksInt}{..}
...
}
}
\usage{
herdsize
}
\description{
A dataset containing info
}
\keyword{datasets}
