% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{am_variables}
\alias{am_variables}
\title{A named list of the variables that go into the common data structure and
their types.}
\usage{
am_variables(varnames = "all")
}
\arguments{
\item{varnames}{if "all" (default), returns all values. Otherwise returns
a subset whose names corresponds to the values given.}
}
\value{
a list as described above.
}
\description{
The names of the list elements returned are the variable names defined in
the common data structure, and the values can be one of three things:
\itemize{
\item Unnamed vector: represents the legal values that that variable can have;
\item Empty vector of a given type: defines the legal type of that variable (
but the actual value is not restricted);
\item Named numeric vector with values between 0 and 1: the names represent the
legal values and the numeric value represent the probability with which
the values should appear when generating artificial data.
}
}
\details{
Note that the legal values given here are  tailored to be used for
generating artificial data using the \code{amu_dummy_data} function, and to
loosely mimic the Swedish cattle sector. The pool of legal values
specified here is therefore not universal; see the file  under
\code{inst/extdata/content_mapping_AMU.xlsx} for the full specification of
the common data structure.
}
