% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_data.R
\name{amu_dummy_data}
\alias{amu_dummy_data}
\title{Generate some random AMU dummy data adhering to the common data structure.
Limited to Sweden and cattle and. Diagnoses are fake (simply a random number)
but ATC codes and medical product IDs are real and fetched from the
"atc.csv" file in this package. The EntryID is merely an index of the row
number. NOTE! This function takes quite some time to run for large values of
\code{n_rows}.}
\usage{
amu_dummy_data(n_rows, start_date, end_date)
}
\arguments{
\item{n_rows}{the number of rows in the dataset}

\item{start_date}{The earliest possible date of transaction}

\item{end_date}{The latest possible date of transaction.}
}
\value{
a data.frame with the dummy data
}
\description{
Generate some random AMU dummy data adhering to the common data structure.
Limited to Sweden and cattle and. Diagnoses are fake (simply a random number)
but ATC codes and medical product IDs are real and fetched from the
"atc.csv" file in this package. The EntryID is merely an index of the row
number. NOTE! This function takes quite some time to run for large values of
\code{n_rows}.
}
